requirements.txt
fastapi
uvicorn
python-multipart
httpx
boto3
sqlalchemy
psycopg2-binary
celery
redis
slowapi
google-generativeai>=0.3.0
python-dotenv
Pillow
openai>=1.0.0